+++ GeoTIFF-Parser/GeoTIFF_Parser.h
@@ -18,7 +18,7 @@
 bool LoadGeoTIFF(std::string filePath);
 void UnloadGeoTIFF();
 const Array2D * GetPointerToBitmap();
-double GetSample(int x, int y, int sampleOrder);
+double GetSample(unsigned long int x, unsigned long int y, unsigned int sampleOrder);
 
 void DisplayTIFFDetailsOnCLI();
 void DisplayGeoTIFFDetailsOnCLI();
